update of both
Lots of changes, firstly, I set up the main layout of the profile card. Next, I linked a stylesheet and setup the colors and text fonts. I then proceeded to style the page. I had to utilize transform: translateY():; for the profile picture, then the rest of the information on top of adding margins. Still have to add in the svg images in the background, but I think the card is looking pretty good to see on mobile.