updated index.html and style.css
Firstly, I added the container for the 2 background svg's so that I can modify them accordingly. I then set the main card z-index to 1 to set it on top of the background, which I set to 0. From there, I tweaked the scale and positioning of each background svg element to look decent on mobile. I will upload to github and setup the media queries for the desktop layout at a later session.