updated index and styles
In index, I  played around with how I wanted the profile card name to be oriented with the age and location. As for the styles file, I set the body width and height to the dynamic variant (dvh and dvw). I also continued to style the layout and need to fix the spacing under the profile (avatar) image and above the name/age/location. I also need to apply the font styles, weights and sizing.