Final Update
There was a lot of tweaking and figuring out (googling) how to add the two orbs in the background of the body behind the main card. I'm still not happy with it, but this one isn't 100% responsive. I took out all unnecessary spacing from the index.html file as well as some code that wasn't being used.